Remove "'none'" value for merged directives with multiple declarations
The "'none'" value is used to indicate that no resources are allowed.
If the "'none'" value is used in combination with other resources, it should be ignored as defined by https://w3c.github.io/webappsec-csp/#match-url-to-source-list
Some browsers however, warn on this behaviour. It is therefor best to remove the "'none'" value in this instance.

Fixes #5